# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trading Simulator API",
    "version": "1.0.0",
    "description": "API for the Trading Simulator - a platform for simulated cryptocurrency trading competitions\n      \n## Authentication Guide\n\nThis API uses Bearer token authentication. All protected endpoints require the following header:\n\n- **Authorization**: Bearer your-api-key\n\nWhere \"your-api-key\" is the API key provided during team registration.\n\n### Authentication Examples\n\n**cURL Example:**\n\n```bash\ncurl -X GET \"https://api.example.com/api/account/balances\" \\\n  -H \"Authorization: Bearer abc123def456_ghi789jkl012\" \\\n  -H \"Content-Type: application/json\"\n```\n\n**JavaScript Example:**\n\n```javascript\nconst fetchData = async () => {\n  const apiKey = 'abc123def456_ghi789jkl012';\n  const response = await fetch('https://api.example.com/api/account/balances', {\n    headers: {\n      'Authorization': `Bearer ${apiKey}`,\n      'Content-Type': 'application/json'\n    }\n  });\n  \n  return await response.json();\n};\n```\n\nFor convenience, we provide an API client that handles authentication automatically. See `docs/examples/api-client.ts`.\n      ",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "ISC License",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
...
    },
    {
...
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key provided in the Authorization header using Bearer token authentication"
      }
    },
...
  },
...
  "paths": {
...
    "/api/competition/leaderboard": {
      "get": {
        "tags": [
          "Competition"
        ],
        "summary": "Get competition leaderboard",
        "description": "Get the leaderboard for the active competition or a specific competition",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Bearer token for authentication (format \"Bearer YOUR_API_KEY\")",
            "example": "Bearer abc123def456_ghi789jkl012"
          },
          {
            "in": "query",
            "name": "competitionId",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Optional competition ID (if not provided, the active competition is used)"
          }
        ],
        "responses": {
          "200": {
            "description": "Competition leaderboard",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "competition": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Competition ID"
                        },
                        "name": {
                          "type": "string",
                          "description": "Competition name"
                        },
                        "description": {
                          "type": "string",
                          "nullable": true,
                          "description": "Competition description"
                        },
                        "startDate": {
                          "type": "string",
                          "format": "date-time",
                          "description": "Competition start date"
                        },
                        "endDate": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true,
                          "description": "Competition end date"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "PENDING",
                            "ACTIVE",
                            "COMPLETED"
                          ],
                          "description": "Competition status"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "When the competition was created"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "When the competition was last updated"
                        }
                      }
                    },
                    "leaderboard": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rank": {
                            "type": "integer",
                            "description": "Team rank on the leaderboard"
                          },
                          "teamId": {
                            "type": "string",
                            "description": "Team ID"
                          },
                          "teamName": {
                            "type": "string",
                            "description": "Team name"
                          },
                          "portfolioValue": {
                            "type": "number",
                            "description": "Current portfolio value in USD"
                          },
                          "active": {
                            "type": "boolean",
                            "description": "Whether the team is active"
                          },
                          "deactivationReason": {
                            "type": "string",
                            "nullable": true,
                            "description": "Reason for deactivation if applicable"
                          }
                        }
                      }
                    },
                    "hasInactiveTeams": {
                      "type": "boolean",
                      "description": "Indicates if any teams are inactive"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - No active competition and no competitionId provided"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "403": {
            "description": "Forbidden - Team not participating in the competition"
          },
          "404": {
            "description": "Competition not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/competition/status": {
      "get": {
        "tags": [
          "Competition"
        ],
        "summary": "Get competition status",
        "description": "Get the status of the active competition",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Bearer token for authentication (format \"Bearer YOUR_API_KEY\")",
            "example": "Bearer abc123def456_ghi789jkl012"
          }
        ],
        "responses": {
          "200": {
            "description": "Competition status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "active": {
                      "type": "boolean",
                      "description": "Whether there is an active competition"
                    },
                    "competition": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Competition ID"
                        },
                        "name": {
                          "type": "string",
                          "description": "Competition name"
                        },
                        "description": {
                          "type": "string",
                          "nullable": true,
                          "description": "Competition description"
                        },
                        "startDate": {
                          "type": "string",
                          "format": "date-time",
                          "description": "Competition start date"
                        },
                        "endDate": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true,
                          "description": "Competition end date"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "PENDING",
                            "ACTIVE",
                            "COMPLETED"
                          ],
                          "description": "Competition status"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "When the competition was created"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "When the competition was last updated"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "Additional information about the competition status",
                      "nullable": true
                    },
                    "participating": {
                      "type": "boolean",
                      "description": "Whether the authenticated team is participating in the competition",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/competition/rules": {
      "get": {
        "tags": [
          "Competition"
        ],
        "summary": "Get competition rules",
        "description": "Get the rules, rate limits, and other configuration details for the competition",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Bearer token for authentication (format \"Bearer YOUR_API_KEY\")",
            "example": "Bearer abc123def456_ghi789jkl012"
          }
        ],
        "responses": {
          "200": {
            "description": "Competition rules retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "rules": {
                      "type": "object",
                      "properties": {
                        "tradingRules": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "List of trading rules for the competition"
                        },
                        "rateLimits": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "Rate limits for API endpoints"
                        },
                        "availableChains": {
                          "type": "object",
                          "properties": {
                            "svm": {
                              "type": "boolean",
                              "description": "Whether Solana (SVM) is available"
                            },
                            "evm": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              },
                              "description": "List of available EVM chains"
                            }
                          }
                        },
                        "slippageFormula": {
                          "type": "string",
                          "description": "Formula used for calculating slippage"
                        },
                        "portfolioSnapshots": {
                          "type": "object",
                          "properties": {
                            "interval": {
                              "type": "string",
                              "description": "Interval between portfolio snapshots"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - No active competition"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "403": {
            "description": "Forbidden - Team not participating in the competition"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}